go/materialize: update references
Updates the references in `go/materialize` to fix the compilation errors.

This includes lifting `MaterializeOptions` & friends into
`transactions_stream.go`, out of the boilerplate file.
