+++ go/materialize/logging_test.go
@@ -1,4 +1,4 @@
-package boilerplate
+package materialize
 
 import (
 	"sync"
