+++ go/materialize/schedule_config.go
@@ -1,4 +1,4 @@
-package boilerplate
+package materialize
 
 import (
 	"fmt"
