+++ go/materialize/schedule_config_test.go
@@ -1,4 +1,4 @@
-package boilerplate
+package materialize
 
 import (
 	"encoding/json"
